Project Home Page
Our project will be totally based on the GUI which will be distributed among different modules. On the front page there will be the Sentiment Analysis on Roman Urdu Dataset along with the detail about the Developers. Moreover which types of modules have been used for creating the home page will vbe mentioned there itself. And from the home page itself we can jump to another page to analyse the whole project.